+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Fix incorrect format for decimals/doubles ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/88)).
+- Add support for Number Expression in SOQL Geolocation Literal ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/90)).
 
 # 1.0.0-rc.5